requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.12.5
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1606860386887/work
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2 @ file:///C:/ci/psycopg2_1608147659101/work
python-decouple==3.3
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
whitenoise==5.2.0
wincertstore==0.2
